requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
stripe
openai
google-generativeai
email-validator
dnspython
anthropic
chromadb
pypdf
